Method Inactive

+++ ProjectOverdrive.API/ProjectOverdrive.API/Repository/CompanyRepository.cs
@@ -6,6 +6,7 @@ using ProjectOverdrive.API.Data.ValueObjects.Response;
 using ProjectOverdrive.API.Models;
 using ProjectOverdrive.API.Repository.Interfaces;
 using System.Collections.Generic;
+using System.Net;
 
 namespace ProjectOverdrive.API.Repository
 {
@@ -19,33 +20,33 @@ namespace ProjectOverdrive.API.Repository
             _mapper = mapper;
         }
 
-        public async Task<List<CompanyResponse>> SearchCompany()
+        public async Task<List<SearchCompanyResponse>> SearchCompany()
         {
             List<Company> company = await _dbContext.Company
                 .Include(a => a.Address)
                 .ToListAsync();
 
-             return _mapper.Map<List<CompanyResponse>>(company); 
+             return _mapper.Map<List<SearchCompanyResponse>>(company); 
         }
 
-        public async Task<CompanyResponse> SearchCompanyByCnpj(string cnpj)
+        public async Task<SearchCompanyResponse> SearchCompanyByCnpj(string cnpj)
         {
             Company company = await _dbContext.Company
                 .Where(c => c.Cnpj == cnpj)
                 .Include(a => a.Address)
                 .FirstOrDefaultAsync();
 
-            return _mapper.Map<CompanyResponse>(company);
+            return _mapper.Map<SearchCompanyResponse>(company);
         }
 
-        public async Task<CompanyResponse> SearchCompanyByName(string name)
+        public async Task<SearchCompanyResponse> SearchCompanyByName(string name)
         {
             Company company = await _dbContext.Company
                 .Where(c => c.CompanyName == name)
                 .Include(a => a.Address)
                 .FirstOrDefaultAsync();
 
-            return _mapper.Map<CompanyResponse>(company);
+            return _mapper.Map<SearchCompanyResponse>(company);
         }
 
         public async Task<CompanyResponse> SearchPeopleInCompany(int id)
@@ -86,7 +87,7 @@ namespace ProjectOverdrive.API.Repository
                 await _dbContext.Company.AddAsync(company);
                 await _dbContext.SaveChangesAsync();
 
-                return _mapper.Map<CompanyRequest>(company); ;
+                return _mapper.Map<CompanyRequest>(company); 
             }
             else
             {
@@ -124,25 +125,55 @@ namespace ProjectOverdrive.API.Repository
             }  
         }
 
-        public async Task<bool> DeleteCompany(int id)
+        public async Task<CompanyUpdateRequest> InactiveCompany(int id)
         {
-            try
+            Company company = await _dbContext.Company.Where(c => c.Id == id)
+                  .FirstOrDefaultAsync() ?? new Company();
+
+            if (company == null) throw new Exception("Essa empresa não existe no banco de dados");
+
+            var checkPeople = await _dbContext.People
+                .Where(p => p.IdCompany == id)
+                .FirstOrDefaultAsync();
+
+            if (checkPeople == null)
             {
+                company.Status = Enum.Status.Inactive;
+                _dbContext.Company.Update(company);
+                await _dbContext.SaveChangesAsync();
+                return _mapper.Map<CompanyUpdateRequest>(company);
+            }
+            else
+            {
+                throw new Exception("Impossivel inativar, existem pessoas nessa empresa.");
+            }
+
+        }
+
+            public async Task<bool> DeleteCompany(int id)
+        {  
                 Company company = await _dbContext.Company.Where(c => c.Id == id)
-                    .FirstOrDefaultAsync() ?? new Company();
+                   .FirstOrDefaultAsync() ?? new Company();
                 Address address = await _dbContext.Address
                     .Where(a => a.Id == company.IdAddress)
                     .FirstOrDefaultAsync();
                 if (company == null) return false;
-                _dbContext.Company.Remove(company);
-                _dbContext.Address.Remove(address);
-                await _dbContext.SaveChangesAsync();
-                return true;
-            }
-            catch(Exception)
-            {
-                return false;
-            }
+
+                var checkPeople = await _dbContext.People
+                    .Where(p => p.IdCompany == id)
+                    .FirstOrDefaultAsync();
+
+                if (checkPeople == null)
+                {
+                    _dbContext.Company.Remove(company);
+                    _dbContext.Address.Remove(address);
+                    await _dbContext.SaveChangesAsync();
+                    return true;
+                }
+                else
+                {
+                    throw new Exception("Impossivel excluir, existem pessoas nessa empresa.");
+                }
         }   
     }
 }
